+++ stream_sim/__init__.py
@@ -1,3 +1,2 @@
 #!/usr/bin/env python
 """ Nothing to see here. """
-__author__ = "Alex Drlica-Wagner"
