+++ stream_sim/samplers.py
@@ -2,7 +2,6 @@
 """
 Probabilistic samplers.
 """
-__author__ = "Alex Drlica-Wagner"
 
 import numpy as np
 import scipy.stats
